Release Telux_master_v1.56.0
telux: Add an entry in revision history table for 'AG' revision
telux: Support for dynamic configuration of SDK Parameters.
tel: Temp Fix for undefined reference for log API
telux: Compile apps based on library availability
data: Host and Guest VM TelSDK Data Interface Support
power: Validate NAOIP trigger text for VM
audio: add application for incall pcm playback and capture
therm: Addition of APIs for thermal notifications
audio: updated documentation for BT usecases
data: api added to request dds switch
power: telsdk power reference application update for hypervisor framework
power: Power test app update for Hypervisor Framework
power: Updates to power management APIs for hypervisor framework
Revert "wlan: Added changes Post CR Fixes"
Update latest symbols files
wlan: Added changes Post CR Fixes
crypto: add SSR support in crypto accelerator console app
qits: Add support for some config parameters in qits
cv2x: qits support min log
audio: fix deadlock in deletestream in console_app
cv2x: print file header to tx report csv file
crypto: add interface for crypto accelerator SSR
audio: update copyright years
tel: Added CallManager readiness in CallMenu
tel: Output UNAVAILABLE string for invalid values of neighbor cell
data: update onHwAccelerationChanged documentation
cv2x: Disable some cv2x sample apps compilations
tel: Updated ecall reference app to use source and sink audio devices
tel: Update documentation and call-flows related to Call Management
audio: updated sample app and documentation for voice call
power: Introduce VM flag
telux: Fix the incorrect subsystem ready time issue
telux: Fix KW issues.
crypto: label result explicitly
crypto: miscellaneous fixes and updates
audio: rectify transcoder creation
audio: Updated sampling rate info and voice menu
audio: fix for race condition during SSR in console_app

CRs-Fixed: 3427747, 3280503, 3402729, 3406868, 3419308, 3280503, 3380984, 3416224, 3404381, 3395846, 3395741, 3391621, 3385357, 3211498, 3366978, 3380984, 3390487, 3425297, 3425310, 3405576, 3386853, 3396897, 3386853, 3359111, 3236523, 3434227, 3421455, 3390831, 3396992, 3437724

+++ apps/reference/rits/src/qCoder/include/wsmp.h
@@ -25,9 +25,9 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-/*
+ *
  * Changes from Qualcomm Innovation Center are provided under the following license:
+ *
  * Copyright (c) 2022 Qualcomm Innovation Center, Inc. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause-Clear
  */
@@ -244,7 +244,7 @@ typedef struct {
     uint32_t psid;
 
     uint8_t *chan_load_ptr;
-    uint8_t chan_load_len;
+    uint32_t chan_load_len;
     char *payload;      /* payload pointer after wsmp header */
     int payload_len;    /* payload length after wsmp header */
